Set app logger level to DEBUG
If the logger level is unset, Python will set it to the highest
ancestor's logger level, which is ERROR. There was no logging in
production before this commit because all DEBUG and INFO logs were not
passed to MongoHandler.

If a handler's logging level is unset, on the other hand, all log
records are processed.